refactor(Finished): 의미없는 double형 int형으로 변경

+++ src/main/java/blackjack/domain/state/Finished.java
@@ -6,10 +6,10 @@ import blackjack.domain.card.Hand;
 public abstract class Finished extends Started {
 
     protected final static double BLACKJACK_WIN_RATE = 1.5;
-    protected final static double WIN_RATE = 1.0;
-    protected final static double BLACKJACK_TIE_RATE = 1.0;
-    protected final static double TIE_RATE = 0;
-    protected final static double LOSE_RATE = -1.0;
+    protected final static int WIN_RATE = 1;
+    protected final static int BLACKJACK_TIE_RATE = 1;
+    protected final static int TIE_RATE = 0;
+    protected final static int LOSE_RATE = -1;
 
     protected Finished(Hand hand) {
         super(hand);
